C and TS
I changed some c files.

I wrote marcus' ipconfig thing in c so that it could be made into
an exe so that when it was placed within the %PATH%, run could
run the resulting exe without the extension, unlike the previous
vbs file that I had given him.

no.c is just `yes n`. It's bad.

I experimented with the syntax sugar that is c arrays. Weird
stuff.

The d files are just updated to use three space indent levels.

Option.ts has been changed to include better documentation of
ignored parameters for specific Some values.

monkey-count.zig was a codewars problem but I tried to do it in
zig. I don't remember if it works.