Terrible Code TBH
`palindrome.c` is a terrible way to test if a string is a palindrome.
Don't get me wrong. It's not the algorithm. It's everything else.

Removed some assembly output.

`sizeof.c` is there since I never actually know what size integer
types seem to be in C. Hopefully I'll never have to look it up again.

I've just noticed that there's inconsistent naming between the two
solutions for problem four of Project Euler. That's annoying but I've
typed too much in this commit message to undo it. I'm sure there's a
way to change that using git magic but I'm no git magician.