Mostly school related things
I was going to do some stuff with F# for my math class as if it
wasn't confusing enough.
The ipython notebooks were also for math but I don't want em.
I'll bean em next time.
Semver was for @digitalagedragon's compilation issues with TS. Forgot to delete
it. I'll get it next commit.
There's a bit of VB in there too.