optimized_request.ts
I took it out of ts-steam-webapi because it really doesn't belong
in there. I'm slightly sad that I didn't get to use the cool trick
of reading ISIZE from the deflate to allocate the right amount of
memory for the uncompressed content. I suppose gunzipSync probably
does that anyways but it's stealing my glory.