bit_cast
reinterpret_cast, our relationship was long and wonderful. We shared
many intimate moments together but in the end, you refused to let me
cast a pointer to a member function. reinterpret_cast, I know you
care about my safety, but frankly, it's not up to you to decide what
I do. This is *my* foot and I will point *my* gun at it.